refactor: split connector rollout handler into manual/internal endpoints (#16050)
