+++ airbyte-commons-server/src/test/kotlin/io/airbyte/commons/server/handlers/ConnectorRolloutHandlerTest.kt
@@ -5,37 +5,23 @@
 package io.airbyte.commons.server.handlers
 
 import io.airbyte.api.model.generated.ConnectorRolloutFinalizeRequestBody
-import io.airbyte.api.model.generated.ConnectorRolloutManualFinalizeRequestBody
-import io.airbyte.api.model.generated.ConnectorRolloutManualRolloutRequestBody
-import io.airbyte.api.model.generated.ConnectorRolloutManualStartRequestBody
-import io.airbyte.api.model.generated.ConnectorRolloutRead
 import io.airbyte.api.model.generated.ConnectorRolloutRequestBody
 import io.airbyte.api.model.generated.ConnectorRolloutStartRequestBody
-import io.airbyte.api.model.generated.ConnectorRolloutState
 import io.airbyte.api.model.generated.ConnectorRolloutStateTerminal
 import io.airbyte.api.model.generated.ConnectorRolloutStrategy
-import io.airbyte.api.model.generated.ConnectorRolloutTierFilter
-import io.airbyte.api.model.generated.ConnectorRolloutUpdateStateRequestBody
 import io.airbyte.api.problems.throwable.generated.ConnectorRolloutInvalidRequestProblem
 import io.airbyte.api.problems.throwable.generated.ConnectorRolloutMaximumRolloutPercentageReachedProblem
 import io.airbyte.api.problems.throwable.generated.ConnectorRolloutNotEnoughActorsProblem
+import io.airbyte.commons.server.handlers.helpers.ConnectorRolloutHelper
 import io.airbyte.config.ActorDefinitionVersion
-import io.airbyte.config.AttributeName
 import io.airbyte.config.ConnectorEnumRolloutState
 import io.airbyte.config.ConnectorEnumRolloutStrategy
 import io.airbyte.config.ConnectorRollout
-import io.airbyte.config.ConnectorRolloutFilters
 import io.airbyte.config.ConnectorRolloutFinalState
-import io.airbyte.config.CustomerTier
-import io.airbyte.config.CustomerTierFilter
-import io.airbyte.config.Operator
-import io.airbyte.config.User
 import io.airbyte.config.persistence.UserPersistence
-import io.airbyte.connector.rollout.client.ConnectorRolloutClient
 import io.airbyte.connector.rollout.shared.ActorSelectionInfo
 import io.airbyte.connector.rollout.shared.ActorSyncJobInfo
 import io.airbyte.connector.rollout.shared.RolloutActorFinder
-import io.airbyte.connector.rollout.shared.models.ConnectorRolloutOutput
 import io.airbyte.data.helpers.ActorDefinitionVersionUpdater
 import io.airbyte.data.services.ActorDefinitionService
 import io.airbyte.data.services.ConnectorRolloutService
@@ -45,11 +31,9 @@ import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
-import io.mockk.slot
 import io.mockk.verify
 import io.mockk.verifyAll
 import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -68,19 +52,22 @@ internal class ConnectorRolloutHandlerTest {
   private val actorDefinitionVersionUpdater = mockk<ActorDefinitionVersionUpdater>()
   private val scopedConfigurationService = mockk<ScopedConfigurationService>()
   private val userPersistence = mockk<UserPersistence>()
-  private val connectorRolloutClient = mockk<ConnectorRolloutClient>()
   private val rolloutActorFinder = mockk<RolloutActorFinder>()
+  private val connectorRolloutHelper =
+    ConnectorRolloutHelper(
+      connectorRolloutService,
+      actorDefinitionService,
+      userPersistence,
+      rolloutActorFinder,
+    )
   private val connectorRolloutHandler =
     ConnectorRolloutHandler(
-      1,
-      1,
-      1,
       connectorRolloutService,
       actorDefinitionService,
       actorDefinitionVersionUpdater,
-      connectorRolloutClient,
       userPersistence,
       rolloutActorFinder,
+      connectorRolloutHelper,
     )
 
   companion object {
@@ -88,20 +75,6 @@ internal class ConnectorRolloutHandlerTest {
     const val DOCKER_IMAGE_TAG = "0.1"
     val ACTOR_DEFINITION_ID: UUID = UUID.randomUUID()
     val RELEASE_CANDIDATE_VERSION_ID: UUID = UUID.randomUUID()
-    val UPDATED_BY = UUID.randomUUID()
-    val DEFAULT_FILTERS =
-      ConnectorRolloutFilters(
-        customerTierFilters =
-          listOf(
-            CustomerTierFilter(name = AttributeName.TIER, operator = Operator.IN, value = listOf(CustomerTier.TIER_2)),
-          ),
-      )
-
-    @JvmStatic
-    fun validInsertStates() = listOf(ConnectorEnumRolloutState.CANCELED)
-
-    @JvmStatic
-    fun invalidInsertStates() = ConnectorEnumRolloutState.entries.filter { it != ConnectorEnumRolloutState.CANCELED }
 
     @JvmStatic
     fun validStartStates() = listOf(ConnectorEnumRolloutState.INITIALIZED)
@@ -138,10 +111,6 @@ internal class ConnectorRolloutHandlerTest {
       }
 
     @JvmStatic
-    fun workflowStartedInProgress() =
-      listOf(ConnectorEnumRolloutState.WORKFLOW_STARTED, ConnectorEnumRolloutState.IN_PROGRESS, ConnectorEnumRolloutState.PAUSED)
-
-    @JvmStatic
     fun provideConnectorRolloutStateTerminalNonCanceled(): List<ConnectorRolloutStateTerminal> =
       listOf(
         ConnectorRolloutStateTerminal.SUCCEEDED,
@@ -166,7 +135,7 @@ internal class ConnectorRolloutHandlerTest {
 
     val rolloutRead = connectorRolloutHandler.getConnectorRollout(rolloutId)
 
-    assertEquals(connectorRolloutHandler.buildConnectorRolloutRead(expectedRollout, true), rolloutRead)
+    assertEquals(connectorRolloutHelper.buildConnectorRolloutRead(expectedRollout, true), rolloutRead)
 
     verify {
       connectorRolloutService.getConnectorRollout(rolloutId)
@@ -410,100 +379,6 @@ internal class ConnectorRolloutHandlerTest {
       finalizedRollout.updatedAt >= initialRollout.updatedAt,
       "updatedAt from finalizedRollout should be more recent than updatedAt from initialRollout",
     )
-  }
-
-  @Test
-  fun `test listConnectorRollouts`() {
-    val mockActorDefinitionVersion = createMockActorDefinitionVersion()
-
-    val expectedRollouts =
-      listOf(
-        createMockConnectorRollout(UUID.randomUUID(), ACTOR_DEFINITION_ID, RELEASE_CANDIDATE_VERSION_ID),
-        createMockConnectorRollout(UUID.randomUUID(), ACTOR_DEFINITION_ID, RELEASE_CANDIDATE_VERSION_ID),
-      )
-
-    every {
-      actorDefinitionService.getActorDefinitionVersion(
-        ACTOR_DEFINITION_ID,
-        DOCKER_IMAGE_TAG,
-      )
-    } returns Optional.of(mockActorDefinitionVersion)
-    every { connectorRolloutService.listConnectorRollouts(ACTOR_DEFINITION_ID, RELEASE_CANDIDATE_VERSION_ID) } returns expectedRollouts
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns mockActorDefinitionVersion
-
-    val rolloutReads = connectorRolloutHandler.listConnectorRollouts(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-
-    assertEquals(expectedRollouts.map { connectorRolloutHandler.buildConnectorRolloutRead(it, false) }, rolloutReads)
-
-    verify {
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-      actorDefinitionService.getActorDefinitionVersion(any())
-      connectorRolloutService.listConnectorRollouts(ACTOR_DEFINITION_ID, RELEASE_CANDIDATE_VERSION_ID)
-    }
-  }
-
-  @Test
-  fun `test listConnectorRolloutsByActorDefinitionVersion`() {
-    val expectedRollouts =
-      listOf(
-        createMockConnectorRollout(UUID.randomUUID(), ACTOR_DEFINITION_ID, RELEASE_CANDIDATE_VERSION_ID),
-        createMockConnectorRollout(UUID.randomUUID(), ACTOR_DEFINITION_ID, UUID.randomUUID()),
-      )
-
-    every { connectorRolloutService.listConnectorRollouts(ACTOR_DEFINITION_ID) } returns expectedRollouts
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-
-    val rolloutReads = connectorRolloutHandler.listConnectorRollouts(ACTOR_DEFINITION_ID)
-
-    assertEquals(expectedRollouts.map { connectorRolloutHandler.buildConnectorRolloutRead(it, false) }, rolloutReads)
-
-    verify {
-      connectorRolloutService.listConnectorRollouts(ACTOR_DEFINITION_ID)
-      actorDefinitionService.getActorDefinitionVersion(any())
-    }
-  }
-
-  @Test
-  fun `test listAllConnectorRollouts`() {
-    val mockActorDefinitionVersion = createMockActorDefinitionVersion()
-
-    val expectedRollouts =
-      listOf(
-        createMockConnectorRollout(UUID.randomUUID(), ACTOR_DEFINITION_ID, RELEASE_CANDIDATE_VERSION_ID),
-        createMockConnectorRollout(UUID.randomUUID(), UUID.randomUUID(), UUID.randomUUID()),
-      )
-
-    every {
-      actorDefinitionService.getActorDefinitionVersion(any(), any())
-    } returns Optional.of(mockActorDefinitionVersion)
-    every { connectorRolloutService.listConnectorRollouts() } returns expectedRollouts
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-
-    val rolloutReads = connectorRolloutHandler.listConnectorRollouts()
-
-    assertEquals(expectedRollouts.map { connectorRolloutHandler.buildConnectorRolloutRead(it, false) }, rolloutReads)
-
-    verify {
-      connectorRolloutService.listConnectorRollouts()
-      actorDefinitionService.getActorDefinitionVersion(any())
-    }
-  }
-
-  @Test
-  fun `test listConnectorRolloutsNoActorDefinitionVersion`() {
-    val actorDefinitionId = UUID.randomUUID()
-    val releaseCandidateVersionId = UUID.randomUUID()
-
-    every { actorDefinitionService.getActorDefinitionVersion(actorDefinitionId, DOCKER_IMAGE_TAG) } returns Optional.empty()
-
-    val rolloutReads = connectorRolloutHandler.listConnectorRollouts(actorDefinitionId, DOCKER_IMAGE_TAG)
-
-    assertEquals(emptyList<ConnectorRolloutRead>(), rolloutReads)
-
-    verify {
-      actorDefinitionService.getActorDefinitionVersion(actorDefinitionId, DOCKER_IMAGE_TAG)
-    }
-    verify(exactly = 0) { connectorRolloutService.listConnectorRollouts(actorDefinitionId, releaseCandidateVersionId) }
   }
 
   @Test
@@ -994,676 +869,6 @@ internal class ConnectorRolloutHandlerTest {
     assertEquals(nFailed, result.values.first().getNumFailed())
   }
 
-  @Test
-  fun `test manualStartConnectorRollout with no filters`() {
-    val rolloutId = UUID.randomUUID()
-    val connectorRolloutWorkflowStart =
-      ConnectorRolloutManualStartRequestBody().apply {
-        dockerRepository = DOCKER_REPOSITORY
-        dockerImageTag = DOCKER_IMAGE_TAG
-        actorDefinitionId = ACTOR_DEFINITION_ID
-        updatedBy = UPDATED_BY
-        rolloutStrategy = ConnectorRolloutStrategy.MANUAL
-        migratePins = false
-      }
-    val connectorRollout = createMockConnectorRollout(rolloutId, rolloutStrategy = null)
-
-    every { connectorRolloutService.listConnectorRollouts(any(), any()) } returns listOf(connectorRollout)
-    every { connectorRolloutClient.startRollout(any(), any()) } just Runs
-    every { connectorRolloutService.getConnectorRollout(rolloutId) } returns connectorRollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-    every {
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-    } returns Optional.of(createMockActorDefinitionVersion())
-    every {
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-    } returns Optional.of(createMockActorDefinitionVersion())
-    val rolloutSlot = slot<ConnectorRollout>()
-    every { connectorRolloutService.writeConnectorRollout(capture(rolloutSlot)) } returns connectorRollout
-    every { userPersistence.getUser(any()) } returns
-      Optional.of(
-        User().apply {
-          userId = UUID.randomUUID()
-          email = ""
-        },
-      )
-
-    val result = connectorRolloutHandler.manualStartConnectorRollout(connectorRolloutWorkflowStart)
-
-    assertEquals(connectorRollout.id, result.id)
-    assertEquals(connectorRollout.rolloutStrategy, ConnectorEnumRolloutStrategy.MANUAL)
-    assertEquals(connectorRollout.updatedBy, UPDATED_BY)
-    val capturedRollout = rolloutSlot.captured
-    assertEquals(DEFAULT_FILTERS, capturedRollout.filters)
-
-    verifyAll {
-      connectorRolloutClient.startRollout(any(), any())
-      connectorRolloutService.getConnectorRollout(rolloutId)
-      connectorRolloutService.listConnectorRollouts(any(), any())
-      actorDefinitionService.getActorDefinitionVersion(any())
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-      connectorRolloutService.writeConnectorRollout(any())
-      userPersistence.getUser(any())
-    }
-  }
-
-  @Test
-  fun `test manualStartConnectorRollout with filters`() {
-    val rolloutId = UUID.randomUUID()
-    val connectorRolloutWorkflowStart =
-      ConnectorRolloutManualStartRequestBody().apply {
-        dockerRepository = DOCKER_REPOSITORY
-        dockerImageTag = DOCKER_IMAGE_TAG
-        actorDefinitionId = ACTOR_DEFINITION_ID
-        updatedBy = UPDATED_BY
-        rolloutStrategy = ConnectorRolloutStrategy.MANUAL
-        migratePins = false
-        filters =
-          io.airbyte.api.model.generated
-            .ConnectorRolloutFilters()
-            .tierFilter(
-              ConnectorRolloutTierFilter().tier(
-                io.airbyte.api.model.generated.CustomerTier._1,
-              ),
-            )
-      }
-    val connectorRollout = createMockConnectorRollout(rolloutId, rolloutStrategy = null)
-
-    every { connectorRolloutService.listConnectorRollouts(any(), any()) } returns listOf(connectorRollout)
-    every { connectorRolloutClient.startRollout(any(), any()) } just Runs
-    every { connectorRolloutService.getConnectorRollout(rolloutId) } returns connectorRollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-    every {
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-    } returns Optional.of(createMockActorDefinitionVersion())
-    every {
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-    } returns Optional.of(createMockActorDefinitionVersion())
-    val rolloutSlot = slot<ConnectorRollout>()
-    every { connectorRolloutService.writeConnectorRollout(capture(rolloutSlot)) } returns connectorRollout
-    every { userPersistence.getUser(any()) } returns
-      Optional.of(
-        User().apply {
-          userId = UUID.randomUUID()
-          email = ""
-        },
-      )
-
-    val result = connectorRolloutHandler.manualStartConnectorRollout(connectorRolloutWorkflowStart)
-
-    assertEquals(connectorRollout.id, result.id)
-    assertEquals(connectorRollout.rolloutStrategy, ConnectorEnumRolloutStrategy.MANUAL)
-    assertEquals(connectorRollout.updatedBy, UPDATED_BY)
-    val capturedRollout = rolloutSlot.captured
-    assertEquals(
-      ConnectorRolloutFilters(
-        customerTierFilters =
-          listOf(
-            CustomerTierFilter(name = AttributeName.TIER, operator = Operator.IN, value = listOf(CustomerTier.TIER_1)),
-          ),
-      ),
-      capturedRollout.filters,
-    )
-
-    verifyAll {
-      connectorRolloutClient.startRollout(any(), any())
-      connectorRolloutService.getConnectorRollout(rolloutId)
-      connectorRolloutService.listConnectorRollouts(any(), any())
-      actorDefinitionService.getActorDefinitionVersion(any())
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-      connectorRolloutService.writeConnectorRollout(any())
-      userPersistence.getUser(any())
-    }
-  }
-
-  @Test
-  fun `test getRolloutStrategyForManualStart`() {
-    assertEquals(connectorRolloutHandler.getRolloutStrategyForManualStart(null), ConnectorEnumRolloutStrategy.MANUAL)
-    assertEquals(connectorRolloutHandler.getRolloutStrategyForManualStart(ConnectorRolloutStrategy.MANUAL), ConnectorEnumRolloutStrategy.MANUAL)
-    assertEquals(connectorRolloutHandler.getRolloutStrategyForManualStart(ConnectorRolloutStrategy.AUTOMATED), ConnectorEnumRolloutStrategy.AUTOMATED)
-  }
-
-  @Test
-  fun `test getRolloutStrategyForManualUpdate`() {
-    assertEquals(connectorRolloutHandler.getRolloutStrategyForManualUpdate(null), ConnectorEnumRolloutStrategy.MANUAL)
-    assertEquals(connectorRolloutHandler.getRolloutStrategyForManualUpdate(ConnectorEnumRolloutStrategy.MANUAL), ConnectorEnumRolloutStrategy.MANUAL)
-    assertEquals(
-      connectorRolloutHandler.getRolloutStrategyForManualUpdate(ConnectorEnumRolloutStrategy.AUTOMATED),
-      ConnectorEnumRolloutStrategy.OVERRIDDEN,
-    )
-  }
-
-  @ParameterizedTest
-  @MethodSource("workflowStartedInProgress")
-  fun `test manualDoConnectorRolloutWorkflowUpdate workflow already in progress`(state: ConnectorEnumRolloutState) {
-    val rolloutId = UUID.randomUUID()
-    val actorIds = listOf(UUID.randomUUID(), UUID.randomUUID())
-    val connectorRolloutWorkflowUpdate =
-      ConnectorRolloutManualRolloutRequestBody().apply {
-        dockerRepository = DOCKER_REPOSITORY
-        dockerImageTag = DOCKER_IMAGE_TAG
-        actorDefinitionId = ACTOR_DEFINITION_ID
-        id = rolloutId
-        this.actorIds = actorIds
-      }
-    val connectorRollout = createMockConnectorRollout(rolloutId)
-    // Rollout has been started
-    connectorRollout.apply { this.state = state }
-
-    every { connectorRolloutClient.doRollout(any(), any()) } just Runs
-    every { connectorRolloutService.getConnectorRollout(rolloutId) } returns connectorRollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-
-    connectorRolloutHandler.manualDoConnectorRollout(connectorRolloutWorkflowUpdate)
-
-    verifyAll {
-      connectorRolloutClient.doRollout(any(), any())
-      connectorRolloutService.getConnectorRollout(rolloutId)
-    }
-
-    // Verify that startWorkflow() was not called because the rollout is already in progress
-    verify(exactly = 0) { connectorRolloutClient.startRollout(any(), any()) }
-  }
-
-  @ParameterizedTest
-  @MethodSource("workflowStartedInProgress")
-  fun `test manualDoConnectorRolloutWorkflowUpdate rollout already running with different filters raises`(state: ConnectorEnumRolloutState) {
-    val rolloutId = UUID.randomUUID()
-    val actorIds = listOf(UUID.randomUUID(), UUID.randomUUID())
-    val connectorRolloutWorkflowUpdate =
-      ConnectorRolloutManualRolloutRequestBody().apply {
-        dockerRepository = DOCKER_REPOSITORY
-        dockerImageTag = DOCKER_IMAGE_TAG
-        actorDefinitionId = ACTOR_DEFINITION_ID
-        id = rolloutId
-        this.actorIds = actorIds
-        this.filters =
-          io.airbyte.api.model.generated
-            .ConnectorRolloutFilters()
-            .tierFilter(
-              ConnectorRolloutTierFilter().tier(
-                io.airbyte.api.model.generated.CustomerTier._1,
-              ),
-            )
-      }
-    val connectorRollout = createMockConnectorRollout(rolloutId)
-    // Rollout has been started
-    connectorRollout.apply { this.state = state }
-    // Rollout filters differ from the input filters
-    connectorRollout.filters = DEFAULT_FILTERS
-
-    every { connectorRolloutClient.doRollout(any(), any()) } just Runs
-    every { connectorRolloutService.getConnectorRollout(rolloutId) } returns connectorRollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-
-    assertThrows<RuntimeException> { connectorRolloutHandler.manualDoConnectorRollout(connectorRolloutWorkflowUpdate) }
-
-    verify(exactly = 0) { connectorRolloutClient.startRollout(any(), any()) }
-    verify(exactly = 0) { connectorRolloutClient.doRollout(any(), any()) }
-  }
-
-  @Test
-  fun `test manualDoConnectorRolloutWorkflowUpdate rollout not started uses default filters`() {
-    val rolloutId = UUID.randomUUID()
-    val actorIds = listOf(UUID.randomUUID(), UUID.randomUUID())
-    val connectorRolloutWorkflowUpdate =
-      ConnectorRolloutManualRolloutRequestBody().apply {
-        dockerRepository = DOCKER_REPOSITORY
-        dockerImageTag = DOCKER_IMAGE_TAG
-        actorDefinitionId = ACTOR_DEFINITION_ID
-        id = rolloutId
-        this.actorIds = actorIds
-      }
-    val connectorRollout = createMockConnectorRollout(rolloutId)
-
-    every { connectorRolloutClient.startRollout(any(), any()) } just Runs
-    every { connectorRolloutClient.doRollout(any(), any()) } just Runs
-    every { connectorRolloutService.getConnectorRollout(rolloutId) } returns connectorRollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-    every { connectorRolloutService.listConnectorRollouts(any(), any()) } returns listOf(connectorRollout)
-    every {
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-    } returns Optional.of(createMockActorDefinitionVersion())
-    every {
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-    } returns Optional.of(createMockActorDefinitionVersion())
-
-    val rolloutSlot = slot<ConnectorRollout>()
-    every { connectorRolloutService.writeConnectorRollout(capture(rolloutSlot)) } returns connectorRollout
-
-    connectorRolloutHandler.manualDoConnectorRollout(connectorRolloutWorkflowUpdate)
-
-    val capturedRollout = rolloutSlot.captured
-    assertEquals(DEFAULT_FILTERS, capturedRollout.filters)
-
-    verifyAll {
-      connectorRolloutService.listConnectorRollouts(any(), any())
-      connectorRolloutClient.startRollout(any(), any())
-      connectorRolloutClient.doRollout(any(), any())
-      connectorRolloutService.getConnectorRollout(rolloutId)
-      actorDefinitionService.getActorDefinitionVersion(any())
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-      connectorRolloutService.writeConnectorRollout(any())
-    }
-  }
-
-  @Test
-  fun `test manualDoConnectorRolloutWorkflowUpdate workflow not started with filters`() {
-    val rolloutId = UUID.randomUUID()
-    val actorIds = listOf(UUID.randomUUID(), UUID.randomUUID())
-    val connectorRolloutWorkflowUpdate =
-      ConnectorRolloutManualRolloutRequestBody().apply {
-        dockerRepository = DOCKER_REPOSITORY
-        dockerImageTag = DOCKER_IMAGE_TAG
-        actorDefinitionId = ACTOR_DEFINITION_ID
-        id = rolloutId
-        this.actorIds = actorIds
-        filters =
-          io.airbyte.api.model.generated
-            .ConnectorRolloutFilters()
-            .tierFilter(
-              ConnectorRolloutTierFilter().tier(
-                io.airbyte.api.model.generated.CustomerTier._1,
-              ),
-            )
-      }
-    val connectorRollout = createMockConnectorRollout(rolloutId)
-    // Rollout has been initialized, but workflow hasn't been started
-    connectorRollout.apply { this.state = ConnectorEnumRolloutState.INITIALIZED }
-    connectorRollout.filters =
-      ConnectorRolloutFilters(
-        customerTierFilters =
-          listOf(
-            CustomerTierFilter(name = AttributeName.TIER, operator = Operator.IN, value = listOf(CustomerTier.TIER_1)),
-          ),
-      )
-
-    every { connectorRolloutClient.startRollout(any(), any()) } just Runs
-    every { connectorRolloutClient.doRollout(any(), any()) } just Runs
-    every { connectorRolloutService.getConnectorRollout(rolloutId) } returns connectorRollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-    every { connectorRolloutService.listConnectorRollouts(any(), any()) } returns listOf(connectorRollout)
-    every {
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-    } returns Optional.of(createMockActorDefinitionVersion())
-    every {
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-    } returns Optional.of(createMockActorDefinitionVersion())
-
-    val rolloutSlot = slot<ConnectorRollout>()
-    every { connectorRolloutService.writeConnectorRollout(capture(rolloutSlot)) } returns connectorRollout
-
-    connectorRolloutHandler.manualDoConnectorRollout(connectorRolloutWorkflowUpdate)
-
-    val capturedRollout = rolloutSlot.captured
-    assertEquals(
-      ConnectorRolloutFilters(
-        customerTierFilters =
-          listOf(
-            CustomerTierFilter(name = AttributeName.TIER, operator = Operator.IN, value = listOf(CustomerTier.TIER_1)),
-          ),
-      ),
-      capturedRollout.filters,
-    )
-
-    verifyAll {
-      connectorRolloutService.listConnectorRollouts(any(), any())
-      connectorRolloutClient.startRollout(any(), any())
-      connectorRolloutClient.doRollout(any(), any())
-      connectorRolloutService.getConnectorRollout(rolloutId)
-      actorDefinitionService.getActorDefinitionVersion(any())
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-      connectorRolloutService.writeConnectorRollout(any())
-    }
-  }
-
-  @ParameterizedTest
-  @MethodSource("validFinalizeStates")
-  fun `test manualFinalizeConnectorRollout`(initialState: ConnectorEnumRolloutState) {
-    val rolloutId = UUID.randomUUID()
-    val rollout = createMockConnectorRollout(rolloutId)
-    rollout.apply { state = initialState }
-    val state = ConnectorRolloutStateTerminal.SUCCEEDED
-    val connectorRolloutFinalizeWorkflowUpdate =
-      ConnectorRolloutManualFinalizeRequestBody().apply {
-        dockerRepository = "airbyte/source-faker"
-        dockerImageTag = "0.1"
-        actorDefinitionId = UUID.randomUUID()
-        id = rolloutId
-        this.state = state
-      }
-
-    every { connectorRolloutService.getConnectorRollout(any()) } returns rollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-    if (initialState == ConnectorEnumRolloutState.INITIALIZED) {
-      every { connectorRolloutClient.startRollout(any(), any()) } just Runs
-    }
-    every { connectorRolloutClient.finalizeRollout(any(), any()) } returns Unit
-
-    connectorRolloutHandler.manualFinalizeConnectorRollout(connectorRolloutFinalizeWorkflowUpdate)
-
-    verifyAll {
-      connectorRolloutService.getConnectorRollout(any())
-      actorDefinitionService.getActorDefinitionVersion(any())
-      if (initialState == ConnectorEnumRolloutState.INITIALIZED) {
-        connectorRolloutClient.startRollout(any(), any())
-      }
-      connectorRolloutClient.finalizeRollout(any(), any())
-    }
-  }
-
-  @ParameterizedTest
-  @MethodSource("validFinalizeStates")
-  fun `test manualFinalizeConnectorRollout with pin retention`(initialState: ConnectorEnumRolloutState) {
-    val rolloutId = UUID.randomUUID()
-    val rollout = createMockConnectorRollout(rolloutId)
-    rollout.apply { state = initialState }
-    val state = ConnectorRolloutStateTerminal.SUCCEEDED
-    val connectorRolloutFinalizeWorkflowUpdate =
-      ConnectorRolloutManualFinalizeRequestBody().apply {
-        dockerRepository = "airbyte/source-faker"
-        dockerImageTag = "0.1"
-        actorDefinitionId = UUID.randomUUID()
-        id = rolloutId
-        this.state = state
-        retainPinsOnCancellation = true
-      }
-
-    every { connectorRolloutService.getConnectorRollout(any()) } returns rollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-    if (initialState == ConnectorEnumRolloutState.INITIALIZED) {
-      every { connectorRolloutClient.startRollout(any(), any()) } just Runs
-    }
-    every { connectorRolloutClient.finalizeRollout(any(), any()) } returns Unit
-
-    connectorRolloutHandler.manualFinalizeConnectorRollout(connectorRolloutFinalizeWorkflowUpdate)
-
-    verifyAll {
-      connectorRolloutService.getConnectorRollout(any())
-      actorDefinitionService.getActorDefinitionVersion(any())
-      if (initialState == ConnectorEnumRolloutState.INITIALIZED) {
-        connectorRolloutClient.startRollout(any(), any())
-      }
-      connectorRolloutClient.finalizeRollout(any(), any())
-    }
-  }
-
-  @ParameterizedTest
-  @MethodSource("workflowStartedInProgress")
-  fun `test manualPauseConnectorRollout`(rolloutState: ConnectorEnumRolloutState) {
-    val rolloutId = UUID.randomUUID()
-    val connectorRolloutWorkflowUpdate =
-      ConnectorRolloutUpdateStateRequestBody().apply {
-        dockerRepository = DOCKER_REPOSITORY
-        dockerImageTag = DOCKER_IMAGE_TAG
-        actorDefinitionId = ACTOR_DEFINITION_ID
-        id = rolloutId
-        state = ConnectorRolloutState.PAUSED
-        pausedReason = "test"
-      }
-    val connectorRollout = createMockConnectorRollout(rolloutId)
-    // Rollout has been started
-    connectorRollout.apply { this.state = state }
-
-    every { connectorRolloutClient.startRollout(any(), any()) } just Runs
-    every { connectorRolloutClient.pauseRollout(any(), any()) } returns ConnectorRolloutOutput(state = ConnectorEnumRolloutState.PAUSED)
-    every { connectorRolloutService.getConnectorRollout(rolloutId) } returns connectorRollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-
-    val result = connectorRolloutHandler.manualPauseConnectorRollout(connectorRolloutWorkflowUpdate)
-
-    assertEquals(connectorRollout.id, result.id)
-    verifyAll {
-      connectorRolloutClient.startRollout(any(), any())
-      connectorRolloutClient.pauseRollout(any(), any())
-      connectorRolloutService.getConnectorRollout(rolloutId)
-      actorDefinitionService.getActorDefinitionVersion(any())
-    }
-  }
-
-  @Test
-  fun `test manualPauseConnectorRollout workflow not started`() {
-    val rolloutId = UUID.randomUUID()
-    val connectorRolloutWorkflowUpdate =
-      ConnectorRolloutUpdateStateRequestBody().apply {
-        dockerRepository = DOCKER_REPOSITORY
-        dockerImageTag = DOCKER_IMAGE_TAG
-        actorDefinitionId = ACTOR_DEFINITION_ID
-        id = rolloutId
-        state = ConnectorRolloutState.PAUSED
-        pausedReason = "test"
-      }
-    val connectorRollout = createMockConnectorRollout(rolloutId)
-    // Rollout has been initialized, but workflow hasn't been started
-    connectorRollout.apply { this.state = ConnectorEnumRolloutState.INITIALIZED }
-
-    every { connectorRolloutClient.startRollout(any(), any()) } just Runs
-    every { connectorRolloutClient.pauseRollout(any(), any()) } returns ConnectorRolloutOutput(state = ConnectorEnumRolloutState.PAUSED)
-    every { connectorRolloutService.getConnectorRollout(rolloutId) } returns connectorRollout
-    every { actorDefinitionService.getActorDefinitionVersion(any()) } returns createMockActorDefinitionVersion()
-
-    val result = connectorRolloutHandler.manualPauseConnectorRollout(connectorRolloutWorkflowUpdate)
-
-    assertEquals(connectorRollout.id, result.id)
-    verifyAll {
-      connectorRolloutClient.startRollout(any(), any())
-      connectorRolloutClient.pauseRollout(any(), any())
-      connectorRolloutService.getConnectorRollout(rolloutId)
-      actorDefinitionService.getActorDefinitionVersion(any())
-    }
-  }
-
-  @Test
-  fun `test getOrCreateAndValidateManualStartInput updates rollout when already exists in INITIALIZED state`() {
-    val rolloutId = UUID.randomUUID()
-    val dockerImageTag = "0.1"
-    val actorDefinitionId = UUID.randomUUID()
-    val actorDefinitionVersion = createMockActorDefinitionVersion()
-
-    val connectorRollout =
-      createMockConnectorRollout(rolloutId).apply {
-        this.state = ConnectorEnumRolloutState.INITIALIZED
-      }
-
-    every { actorDefinitionService.getActorDefinitionVersion(actorDefinitionId, dockerImageTag) } returns Optional.of(actorDefinitionVersion)
-    every { connectorRolloutService.listConnectorRollouts(actorDefinitionId, actorDefinitionVersion.versionId) } returns listOf(connectorRollout)
-    every {
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-    } returns Optional.of(createMockActorDefinitionVersion())
-    every { connectorRolloutService.writeConnectorRollout(any()) } returns connectorRollout
-
-    val result =
-      connectorRolloutHandler.getOrCreateAndValidateManualStartInput(
-        DOCKER_REPOSITORY,
-        actorDefinitionId,
-        dockerImageTag,
-        UPDATED_BY,
-        ConnectorRolloutStrategy.MANUAL,
-        null,
-        null,
-        null,
-      )
-
-    assertEquals(connectorRollout.id, result.id)
-    verifyAll {
-      actorDefinitionService.getActorDefinitionVersion(actorDefinitionId, dockerImageTag)
-      connectorRolloutService.listConnectorRollouts(actorDefinitionId, actorDefinitionVersion.versionId)
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-      connectorRolloutService.writeConnectorRollout(any())
-    }
-  }
-
-  @Test
-  fun `test getOrCreateAndValidateManualStartInput throws when initial version is not found`() {
-    val rolloutId = UUID.randomUUID()
-    val dockerImageTag = "0.1"
-    val actorDefinitionId = UUID.randomUUID()
-    val actorDefinitionVersion = createMockActorDefinitionVersion()
-
-    val connectorRollout =
-      createMockConnectorRollout(rolloutId).apply {
-        this.state = ConnectorEnumRolloutState.INITIALIZED
-      }
-
-    every { actorDefinitionService.getActorDefinitionVersion(actorDefinitionId, dockerImageTag) } returns Optional.of(actorDefinitionVersion)
-    every { connectorRolloutService.listConnectorRollouts(actorDefinitionId, actorDefinitionVersion.versionId) } returns listOf(connectorRollout)
-    every {
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(any())
-    } returns null
-
-    assertThrows<ConnectorRolloutInvalidRequestProblem> {
-      connectorRolloutHandler.getOrCreateAndValidateManualStartInput(
-        DOCKER_REPOSITORY,
-        actorDefinitionId,
-        dockerImageTag,
-        UPDATED_BY,
-        ConnectorRolloutStrategy.MANUAL,
-        null,
-        null,
-        null,
-      )
-    }
-  }
-
-  @Test
-  fun `test getOrCreateAndValidateManualStartInput creates new connector rollout if not found`() {
-    val rolloutId = UUID.randomUUID()
-    val actorDefinitionVersion = createMockActorDefinitionVersion()
-
-    val connectorRollout =
-      createMockConnectorRollout(rolloutId).apply {
-        this.state = ConnectorEnumRolloutState.INITIALIZED
-      }
-
-    every { connectorRolloutService.listConnectorRollouts(ACTOR_DEFINITION_ID, actorDefinitionVersion.versionId) } returns emptyList()
-    every { actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG) } returns Optional.of(actorDefinitionVersion)
-    every { actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(ACTOR_DEFINITION_ID) } returns Optional.of(actorDefinitionVersion)
-    every { connectorRolloutService.writeConnectorRollout(any()) } returns connectorRollout
-
-    connectorRolloutHandler.getOrCreateAndValidateManualStartInput(
-      DOCKER_REPOSITORY,
-      ACTOR_DEFINITION_ID,
-      DOCKER_IMAGE_TAG,
-      UPDATED_BY,
-      ConnectorRolloutStrategy.MANUAL,
-      null,
-      null,
-      null,
-    )
-
-    verifyAll {
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-      connectorRolloutService.listConnectorRollouts(ACTOR_DEFINITION_ID, actorDefinitionVersion.versionId)
-      connectorRolloutService.writeConnectorRollout(any())
-      actorDefinitionService.getDefaultVersionForActorDefinitionIdOptional(ACTOR_DEFINITION_ID)
-    }
-  }
-
-  @Test
-  fun `test getOrCreateAndValidateManualStartInput throws ConnectorRolloutInvalidRequestProblem when actor definition version not found`() {
-    every { actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG) } returns Optional.empty()
-
-    assertThrows<ConnectorRolloutInvalidRequestProblem> {
-      connectorRolloutHandler.getOrCreateAndValidateManualStartInput(
-        DOCKER_REPOSITORY,
-        ACTOR_DEFINITION_ID,
-        DOCKER_IMAGE_TAG,
-        UPDATED_BY,
-        ConnectorRolloutStrategy.MANUAL,
-        null,
-        null,
-        null,
-      )
-    }
-
-    verify { actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG) }
-  }
-
-  @Test
-  fun `test getOrCreateAndValidateManualStartInput throws ConnectorRolloutInvalidRequestProblem when docker repository doesn't match`() {
-    val dockerRepository = "wrong-repo"
-    val actorDefinitionVersion = createMockActorDefinitionVersion()
-
-    every { actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG) } returns Optional.of(actorDefinitionVersion)
-
-    assertThrows<ConnectorRolloutInvalidRequestProblem> {
-      connectorRolloutHandler.getOrCreateAndValidateManualStartInput(
-        dockerRepository,
-        ACTOR_DEFINITION_ID,
-        DOCKER_IMAGE_TAG,
-        UPDATED_BY,
-        ConnectorRolloutStrategy.MANUAL,
-        null,
-        null,
-        null,
-      )
-    }
-
-    verifyAll {
-      actorDefinitionService.getActorDefinitionVersion(ACTOR_DEFINITION_ID, DOCKER_IMAGE_TAG)
-    }
-  }
-
-  @Test
-  fun `returns tag for single tier`() {
-    val rollout = createMockConnectorRollout(UUID.randomUUID())
-    rollout.filters =
-      ConnectorRolloutFilters(
-        customerTierFilters =
-          listOf(
-            CustomerTierFilter(
-              operator = Operator.IN,
-              value = listOf(CustomerTier.TIER_1),
-            ),
-          ),
-      )
-
-    val tag = connectorRolloutHandler.createTagFromFilters(rollout.filters)
-    assertEquals("TIER_1", tag)
-  }
-
-  @Test
-  fun `returns tag for multiple tiers sorted`() {
-    val rollout = createMockConnectorRollout(UUID.randomUUID())
-    rollout.filters =
-      ConnectorRolloutFilters(
-        customerTierFilters =
-          listOf(
-            CustomerTierFilter(
-              operator = Operator.IN,
-              value = listOf(CustomerTier.TIER_2, CustomerTier.TIER_0),
-            ),
-          ),
-      )
-
-    val tag = connectorRolloutHandler.createTagFromFilters(rollout.filters)
-    assertEquals("TIER_0-TIER_2", tag)
-  }
-
-  @Test
-  fun `returns null when tier filter is empty`() {
-    val rollout = createMockConnectorRollout(UUID.randomUUID())
-    rollout.filters =
-      ConnectorRolloutFilters(
-        customerTierFilters = emptyList(),
-      )
-
-    val tag = connectorRolloutHandler.createTagFromFilters(rollout.filters)
-    assertNull(tag)
-  }
-
-  @Test
-  fun `returns null when filters is null`() {
-    val rollout = createMockConnectorRollout(UUID.randomUUID())
-
-    val tag = connectorRolloutHandler.createTagFromFilters(rollout.filters)
-    assertNull(tag)
-  }
-
   private fun createMockConnectorRollout(
     id: UUID,
     actorDefinitionId: UUID = ACTOR_DEFINITION_ID,
@@ -1686,12 +891,6 @@ internal class ConnectorRolloutHandlerTest {
       expiresAt = OffsetDateTime.now().plusDays(1).toEpochSecond(),
       tag = null,
     )
-
-  private fun createMockConnectorRolloutStartRequestBody(): ConnectorRolloutStartRequestBody =
-    ConnectorRolloutStartRequestBody()
-      .id(UUID.randomUUID())
-      .workflowRunId(UUID.randomUUID().toString())
-      .rolloutStrategy(ConnectorRolloutStrategy.MANUAL)
 
   private fun createMockConnectorRolloutRequestBody(
     rolloutId: UUID,
